main/requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
mistralai>=1.2.0
langgraph>=0.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
gradio>=4.44.0
chromadb>=0.5.11
sentence-transformers>=3.0.1
